requirements.txt
aiohttp==3.12.7
pydantic==2.11.5
cacheout==0.16.0
pillow==11.2.1
numpy==2.3.0
customtkinter==5.2.2